+++ mofan/Reinforcement-learning-with-tensorflow/contents/ZMY_DQN_max/DQN_run_this.py
@@ -1,5 +1,5 @@
-import brain
+import DQN_brain
 #2020/10/02
-Mymodel = brain.System_A3C_Model(10000, 200)
+Mymodel = DQN_brain.System_A3C_Model(10000, 200)
 
 ###每个action前先算一下信道总增益h_total，因为要保证在一个action中h_total相等，防止在一个action中多次调用h函数计算
